axf/requirement.txt
Django==2.0.7
django-cors-headers==2.4.0
djangorestframework==3.9.2
pymysql
redis==2.10.6
django-redis==4.10.0
django-redis-cache==1.8.1
django_filter